% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{adjustDepths}
\alias{adjustDepths}
\title{Adjust Bin Depths}
\usage{
adjustDepths(adp)
}
\arguments{
\item{adp}{an oce object contasining adcp data as well as alternate pressure data from another instrument}
}
\description{
if in processing you have inserted pressure from another instrument and
choose to use these new pressure values to calculate bin depths then this
function can be used to adjust bin depths based on more accurate pressure
readings.
}
\details{
Another similar function which can adjust bin depths based on existing
pressure data is \code{\link[oce:binMapAdp]{binMapAdp}}
which can map adp bins to be at consistent depths with pressure.
}
