% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{adpNC}
\alias{adpNC}
\title{NetCDF creation from adp object}
\usage{
adpNC(adp, name)
}
\arguments{
\item{adp}{an adp object}

\item{name}{text string which will name netCDF file}
}
\description{
Creates standardized netCDF file from adp object (produced from
  \code{\link[ADCP:adpCombine]{adpCombine}})

  Standardized name of file can be created with:
  (\code{\link[ADCP:name.file]{name.file}})

  product will meet CF compliance, ERDDAP standards, BODC/SDN, and DFO/MEDS
  standards
}
