% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{binMap}
\alias{binMap}
\title{Bin Map}
\usage{
binMap(obj)
}
\arguments{
\item{obj}{adp object from oce package}
}
\details{
a function to create a bin variable within the adp object which can be mapped onto plots
variable contains list of distance from adp to each bin named by bin number

Once bin variable has been created use
```abline(h = obj[['bin']])```
to create horizontal line denoting bins
}
