% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{binPlot}
\alias{binPlot}
\title{Bin Plot}
\usage{
binPlot(adp, x, path)
}
\arguments{
\item{adp}{an oce adp object}

\item{x}{the matrix of data to be plotted}

\item{path}{file path to which plots will be saved}
}
\value{
a pdf with a series of plots
}
\description{
Bin Plot
}
\details{
Series of plots, separated by depth (bins) of a particular parameter in ADCP data
}
\examples{

binPlot(adp, x = adp[['v']][,,1])
}
