% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyNCmeta.R
\name{copy.nc.meta}
\alias{copy.nc.meta}
\title{Copy metadata from netCDF}
\usage{
copy.nc.meta(ncin, ncout)
}
\arguments{
\item{ncin}{old version netCDF}

\item{ncout}{new netCDF version}
}
\description{
pulls global attributes from netCDF to be put into new format netCDF
}
