% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{endPlots}
\alias{endPlots}
\title{End Plots}
\usage{
endPlots(adpClean, path)
}
\arguments{
\item{adpClean}{an adp object with flags set to NA}

\item{path}{file path to which plots will be saved}
}
\value{
a pdf series of plots including velocity components, pressure and echo intensity
}
\description{
Post Processing summary plots for ADCP data
}
