% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{exportPL}
\alias{exportPL}
\title{Export processing log from adp object to be included in netCDF}
\usage{
exportPL(adp)
}
\arguments{
\item{adp}{an oce adp object of ADCP data}
}
\value{
adp object with processingLog as separate character string in adp[['processing_history']]
}
\description{
This function can be used to export the processing log recorded in the adp
object as a single text string to be included in a netCDF
}
\examples{


adp <- exportPL(adp)

}
