% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{flag}
\alias{flag}
\title{Flag}
\usage{
flag(adp, values)
}
\arguments{
\item{adp}{an adp object from oce}

\item{values}{a list of values indexing the adp object's velocity by time which you wish to flag}
}
\value{
adp object with complete flag array
}
\description{
Flag specific data points within an adp dataset
}
\examples{

adp <- flag(adp, values = list('time' = c(1, 2, 3, 4, 5, 6)))

}
