% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{insertInst}
\alias{insertInst}
\title{Insert Pressure from alternate instrument}
\usage{
insertInst(adp, var, file = adp[["alternate_pressure_file"]],
  offset = adp[["vertical_separation"]])
}
\arguments{
\item{adp}{adp object into which to insert data}

\item{var}{variable you wish to pull from other instrument}

\item{file}{file where data from alternate instrument is stored}

\item{offset}{the mooring separation between the instruments (in metres)}
}
\value{
adp object with data set completed from alternate sources
}
\description{
Use the measured pressure from an alternate instrument positioned on the same
moooring as the ADCP This has benefits of a more accurate pressure reading
which can be translated to depth Note: If the pressure sensor is below the
ADCP the mooring separation distance (offset) should be negative.
}
\details{
Extreme caution should be used when applying this function to anything other
than pressure as there may be significant differences in values despite only
a few metres separation.
}
\examples{

      #list odf files
odflist <- list.files(path = ".", pattern =  "MADCP*...*00.ODF")

      #read list of odf files into adp object
adp <- odf2adp(odflist)



      #add in raw and netCDF metadata and data to adp objecct
raw <- list.files(path = '.', pattern = "*.000")
nc <- list.files(path = '.', pattern = "*.nc")
adp <- adpCombine(adp, raw, nc)


       #insert microcat pressure data
file <- list.files(path = '.', pattern = "MCTD*...*.ODF")
adp <- insertInst(adp, var = 'pressure', file = file)


}
