% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_depth.R
\name{limit_depthbyrmax}
\alias{limit_depthbyrmax}
\title{Limit Depth ADCP PRocessing Step 3.1}
\usage{
limit_depthbyrmax(x, lat = x[["latitude"]])
}
\arguments{
\item{x}{oce object of class adp to be limited}

\item{lat}{latitude of instrument during sampling}
}
\description{
Functions to limit the bounds of ADCP depth data

Can be limited by rmax or time (leaving both for now, one may be deleted in final versions)

Use maximum acceptable range values to determine acceptable depth values
Uses Teledyne RDI equation, Rmax = Dcos(x)
}
