% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R, R/limit_depth.R
\name{limit_depthbytime}
\alias{limit_depthbytime}
\alias{limit_depthbytime}
\title{limit depth by time}
\usage{
limit_depthbytime(adp)

limit_depthbytime(adp)
}
\arguments{
\item{adp}{oce object of class adp to be limited

requires certain meta data features to compute
including pressure, latitude, time, deploymentTime, recoveryTime}

\item{adp}{oce object of class adp to be limited

requires certain meta data features to compute
including pressure, latitude, time, deploymentTime, recoveryTime}
}
\description{
Uses deployment and recovery times to limit depth within times that
insturment was actively and properly sampliing

Uses deployment and recovery times to limit depth within times that
insturment was actively and properly sampliing
}
