% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{limit_time}
\alias{limit_time}
\title{ADCP Processing step 3.4}
\usage{
limit_time(x, tz = "UTC", dt = x[["deploymentTime"]],
  rt = x[["recoveryTime"]])
}
\arguments{
\item{x}{adp object from oce-class adp}

\item{tz}{time zone, default is 'UTC'}

\item{dt}{deployment time of ADCP, default pulls value from metadata}

\item{rt}{recovery time of ADCP, default pulls value from metadata}
}
\value{
adp object with velocities limited to active ADCP measurement times (after deployment and before recovery)
}
\description{
Function limits time from before deployment and after recovery of ADCP
}
