% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameFile.R
\name{name.file}
\alias{name.file}
\title{File naming format}
\usage{
name.file(adp)
}
\arguments{
\item{adp}{an oce class - adp object

required metadata : cruise number, mooring number, serial number, sampling interval!! Make sure all metadata is present when experiencing problems}
}
\description{
pulls metadata from an adp oce object and creates a file name in standard BIO format
}
