% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{odf2adp}
\alias{odf2adp}
\title{ADCP Processing ODF to NetCDF}
\usage{
odf2adp(files, metadata)
}
\arguments{
\item{files}{list of odf files}

\item{metadata}{any extra metadata to be added to net cdf as list form}
}
\description{
Converting individual odf bins to Net cdf standard format
}
\seealso{
Other odf: \code{\link{adpCombine}}
}
