% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{plot_ei}
\alias{plot_ei}
\title{Plot echo intensity}
\usage{
plot_ei(adp, ...)
}
\arguments{
\item{adp}{an oce object with adcp data}
}
\value{
plot of time average echo intensity values by bin with separate beams
}
\description{
Plot echo intensity
}
