% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R, R/plotBin.R
\name{plotBin}
\alias{plotBin}
\alias{plotBin}
\title{Bin by bin plot}
\usage{
plotBin(v)

plotBin(v)
}
\arguments{
\item{v}{variable matrix from adcp data, should be 2 dimensional (time, distance or bin)}
}
\description{
Bin by bin plot

use to plot each "bin" of any chosen variable (u, v, error, echo intensity)
}
