% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{plotQC}
\alias{plotQC}
\title{plotQC}
\usage{
plotQC(obj, QC, ...)
}
\arguments{
\item{obj}{an adp object}

\item{QC}{the paramater you wish to plot,

options are u, v, w, er, ei, pg

*u -> eastward velocity component
*v -> northward velocity component
*w -> upward velocity component
*er -> error velocity
*ei -> echo intensity (first beam only)
*pg -> precent good, sum of 1st and 4th beams}
}
\value{
Plots which show obj (adp) data parameter with flagged values shown
  in red and black lines showing good values
}
\description{
Plots which show flagged vs unflagged values of a variety of parameters, can
be used to visually check quality control and confirm that flagged values are
appropriate
}
