% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{pvPlot}
\alias{pvPlot}
\title{Progressive vector plot}
\usage{
pvPlot(x, control, xlim = c(range(x[["distance"]])),
  ylim = c(range(x[["distance"]])))
}
\arguments{
\item{x}{adp object from oce}

\item{control}{list() with optional object 'bin' which can be used to specify the bins you wish to plot}

\item{xlim}{limiting values for x axis}

\item{ylim}{limiting values for y axis}
}
\value{
plot of progressive vectors,
}
\description{
Progressive vector plot
}
\examples{
pvPlot(adp, control = list('bin' = c(1, 5, 7)))
pvPlot(adp, control = list('bin' = c(1:length(adp[['distance']]))))
}
