% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{qcPlots}
\alias{qcPlots}
\title{Quality Control Plots}
\usage{
qcPlots(adp, QC, path)
}
\arguments{
\item{adp}{an oce adp object}

\item{QC}{the QC parameter you want to inspect, options are listed in details}

\item{path}{file path to which plots will be saved}
}
\value{
a pdf compilation of plots
}
\description{
Quality Control Plots
}
\details{
These quality control plots show a comparison between valid and invalid data.
Invalid data is determined by the flags within the adp object (which will
have been created based on processing procedures). The invalid data is
higlighted in red for each bin and a total is shown at the bottom which gives
the user a percentage of invalid data (or data flagged), this can allow a
user to see if there are particular bins which should be ommitted from the
final export data set.

options for quality control parameters are u, v, w, er, ei, pg

*u -> eastward velocity component
*v -> northward velocity component
*w -> upward velocity component
*er -> error velocity
*ei -> echo intensity (first beam only)
*pg -> precent good, sum of 1st and 4th beams
}
