% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R, R/readAdpEasy.R
\name{read.adp.easy}
\alias{read.adp.easy}
\alias{read.adp.easy}
\title{ADCP PRocessing step 2.0}
\usage{
read.adp.easy(file, metadata)

read.adp.easy(file, metadata)
}
\arguments{
\item{file}{raw ADCP file (.000 format)}

\item{metadata}{csv metadata file from template

returns an object of class adp (from oce package)
uses \code{\link[MASS:oce]{read.adp}}}

\item{file}{raw ADCP file (.000 format)}

\item{metadata}{csv metadata file from template}
}
\description{
Load adp data into R with list that includes all metadata from mooring sheets

Load adp data into R with list that includes all metadata from mooring sheets
}
