% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{read.meta}
\alias{read.meta}
\title{ADCP process step 2.1}
\usage{
read.meta(file, obj)
}
\arguments{
\item{file}{csv file name}

\item{obj}{adp oce object to assign metadata to}
}
\description{
ADCP process step 2.1
}
