% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{startPlots}
\alias{startPlots}
\title{Start Plots}
\usage{
startPlots(adp, path)
}
\arguments{
\item{adp}{an oce adp object}

\item{path}{file path to which plots will be saved}
}
\value{
a pdf series of plots
}
\description{
Plots give a first visualization of ADCP data set
}
\details{
Includes each velocity component as well as pressure over time
}
