% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{adpCombine}
\alias{adpCombine}
\title{ADCP Combine}
\usage{
adpCombine(adp, raw, ncin)
}
\arguments{
\item{adp}{an adp object sourced from ODF files using
\code{\link[ADCP:odf2adp]{odf2adp}}}

\item{raw}{a raw ADCP file (.000)}

\item{ncin}{an archived netCDF file (.nc)}
}
\description{
Combines archived data and metadata from raw, odf and netCDF
 sources into adp object which can be exported as netCDF or saved
}
\details{
\bold{a)	Raw file (.000)}

Contains:


\bold{Data:} heading, pressure, sound speed data, BEAM_02-_04, PGDP_02-_04,
PTCH, ROLL, HGHT, Tx, D( D is not actually equal to depth values recorded in
raw file but calculated using swDepth from recorded pressure values)


\bold{ Metadata: } firmwareVersion, frequency, beamPattern, orientation, beamAngle,
numberOfBeams (janus), pingsPerEnsemble, velocityResolution (pred_accuracy),
lowCorrThresh (valid_correlation_range), percentGdMinimum
(minmax_percent_good), errorVelocityMaximum (error_velocity_threshold),
xmitPulseLength (transmit_pulse_length), falseTargetThresh
(false_target_reject_values), serialNumber, instrumentType(data_type),
bin1Distance


b)	\bold{Archived netCDF}

Contains:


     \bold{ Data:} BEAM_02-_04, PGDP_02-_04, PTCH, ROLL, HGHT, Tx, D
***note this DATA is pulled from the RAW file in case NC file is corrupted or missing


\bold{Metadata: }creation_date,  start_time, stop_time, inst_type, history,
starting_water_layer, ending_water_layer, depth_note, transform,
data_subtype, coord_system, water_mass, pos_const, depth_const, drifter,
var_fill, experiment, project, descript, data_cmnt, fill_flag, composite,
magnetic_variation, , delta_t_sec, time_between_ping_groups, depth:
xducer_offset_from_bottom, depth: bin_size



\bold{c)	Archived ODFs (series for each mooring)}

Contains:

     \bold{Data:} PROCESSED EWCT, NSCT, VCSP, ERRV, BEAM_01 (!which is
     actually the average echo intensity), PGDP_04, time, distance


    \bold{ Metadata:}  units (v, distance), cellSize, numberOfBeams, orientation,
     model, type, serialNumber, ship (platform), scientist
     (chief_scientist), institute (data_origin), cruise (cruise_name), station (mooring),
     countryInstituteCode, cruiseNumber, startTime, latitude, longitude,
     waterDepth, sounding
}
\seealso{
Other odf: \code{\link{odf2adp}}

Other NC: \code{\link{adpNC}}, \code{\link{oceNc_create}}

Other processing: \code{\link{limit_depthbytime}},
  \code{\link{limit_time}}, \code{\link{oceNc_create}}
}
