% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{adpFlag}
\alias{adpFlag}
\title{ADCP processing Step 3.5}
\usage{
adpFlag(adp, pg = adp[["percentgd_threshold"]],
  er = adp[["error_threshold"]])
}
\arguments{
\item{adp, }{an adp object, oce-class}

\item{pg, }{The minimum percent good for evaluating beams one + four of the
adcp (BIO standard is 25)}

\item{er, }{Maximum error velocity for evaluating adcp data (BIO standard is
0.46)}

\item{flagScheme, }{scheme of flags that will be followed, BODC, MEDS, etc}
}
\description{
ADCP processing Step 3.5
}
