% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcFlag.R
\name{adpFlag}
\alias{adpFlag}
\title{ADCP processing Step 3.8}
\usage{
adpFlag(adp, metadata, pg, er)
}
\arguments{
\item{adp, }{an adp object, oce-class}

\item{metadata, }{a list of metadata name = 'value'}

\item{pg, }{The minimum percent good for evaluating beams one + four of the adcp (BIO standard is 25)}

\item{er, }{Maximum error velocity for evaluating adcp data (BIO standard is 0.46)}
}
\details{
adpFlag function
flag an adp object based on a series of parameters including percent good, and error velocity.

This function also defaults to flagging depth values based on the Teledyne RDI standard

Rmax = Dcosx

where Rmax is the maximum acceptable distance range from the ADCP, D is total depth and x is the beam angle of the ADCP.
}
