% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{applyMagneticDeclinationAdp}
\alias{applyMagneticDeclinationAdp}
\title{ADCP Processing step 3.2}
\usage{
applyMagneticDeclinationAdp(x, lat = x[["latitude"]],
  lon = x[["longitude"]], st = x[["deploymentTime"]],
  et = x[["recoveryTime"]], tz = "UTC", type = "average")
}
\arguments{
\item{x}{adp object from oce-class}

\item{lat}{latitude}

\item{lon}{longitude}

\item{st}{start/ deployment time}

\item{et}{end/ recovery time}

\item{type}{average or interpolated}
}
\value{
adp object with magnetic declination applied given coordinates

If type == average, an average is taken from start and end time declinations and applied uniformly

If type == interpolated, the rate of declination is used over time series
}
\description{
apply magnetic declination to ADCP data
uses \code{\link[oce:magneticField]{magneticField}} to calculate declination and
\code{\link[oce:enuToOther]{enuToOTher}} to apply variation to data set
}
