% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{insertInst}
\alias{insertInst}
\title{Title}
\usage{
insertInst(adp, var, file, offset = 0)
}
\arguments{
\item{adp}{adp object into which to insert data}

\item{var}{variable you wish to pull from other instrument}

\item{file}{file where data from alternate instrument is stored}

\item{offset}{the mooring separation between the instruments (in metres)}
}
\value{
adp object with data set completed from alternate sources
}
\description{
Title
}
\examples{

      #list odf files
odflist <- list.files(path = ".", pattern =  "MADCP*...*00.ODF")

      #read list of odf files into adp object
adp <- odf2adp(odflist)



      #add in raw and netCDF metadata and data to adp objecct
raw <- list.files(path = '.', pattern = "*.000")
nc <- list.files(path = '.', pattern = "*.nc")
adp <- adpCombine(adp, raw, nc)


       #insert microcat pressure data
file <- list.files(path = '.', pattern = "MCTD*...*.ODF")
adp <- insertInst(adp, var = 'pressure', file = file)


}
