% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{limit_depthbyrmax}
\alias{limit_depthbyrmax}
\title{ADCP Processing step 3.3 Limit depth by rmax}
\usage{
limit_depthbyrmax(x, lat = x[["latitude"]])
}
\arguments{
\item{x}{oce object of class adp to be limited}

\item{lat}{latitude of instrument during sampling}
}
\description{
ADCP Processing step 3.3 Limit depth by rmax
}
