% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R, R/limit_depth.R
\name{limit_depthbyrmax}
\alias{limit_depthbyrmax}
\alias{limit_depthbyrmax}
\title{ADCP Processing step 3.3
Limit depth by rmax}
\usage{
limit_depthbyrmax(x, lat = x[["latitude"]])

limit_depthbyrmax(x, lat = x[["latitude"]])
}
\arguments{
\item{x}{oce object of class adp to be limited}

\item{lat}{latitude of instrument during sampling}

\item{x}{oce object of class adp to be limited}

\item{lat}{latitude of instrument during sampling}
}
\description{
Use maximum acceptable range values to determine acceptable depth values
Uses Teledyne RDI equation, Rmax = Dcos(x)

Functions to limit the bounds of ADCP depth data

Can be limited by rmax or time (leaving both for now, one may be deleted in final versions)

Use maximum acceptable range values to determine acceptable depth values
Uses Teledyne RDI equation, Rmax = Dcos(x)
}
