% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{limit_depthbytime}
\alias{limit_depthbytime}
\title{limit depth by time}
\usage{
limit_depthbytime(adp, tz = "UTC")
}
\arguments{
\item{adp}{oce object of class adp to be limited

 requires certain meta data features to compute including pressure, latitude,
 time, deployment_time, recovery_time}
}
\description{
Uses deployment and recovery times to limit depth within times
 that insturment was actively and properly sampliing
}
