% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{limit_time}
\alias{limit_time}
\title{ADCP Processing step 3.4}
\usage{
limit_time(x, tz = "UTC", dt = x[["time_coverage_start"]],
  rt = x[["time_coverage_end"]])
}
\arguments{
\item{x}{adp object from oce-class adp}

\item{tz}{time zone, default is 'UTC'}

\item{dt}{deployment time of ADCP, default pulls value from metadata (time_coverage_start)}

\item{rt}{recovery time of ADCP, default pulls value from metadata (time_coverage_end)}
}
\value{
adp object with velocities limited to active ADCP measurement times
 (after deployment and before recovery)
}
\description{
Function limits variable (time, salinity, pressure, temperature,
 pitch, roll, heading) from before deployment and after recovery of ADCP
}
\seealso{
Other processing: \code{\link{adpCombine}},
  \code{\link{limit_depthbytime}},
  \code{\link{oceNc_create}}
}
