% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBin.R
\name{plotBin}
\alias{plotBin}
\title{use to plot each "bin" of any chosen variable (u, v, error, echo intensity)}
\usage{
plotBin(v)
}
\description{
use to plot each "bin" of any chosen variable (u, v, error, echo intensity)
}
