% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R
\name{read.adp.easy}
\alias{read.adp.easy}
\title{ADCP Processing step 2.0}
\usage{
read.adp.easy(file, metadata)
}
\arguments{
\item{file}{raw ADCP file (.000 format)}

\item{metadata}{csv metadata file from template}
}
\description{
ADCP Processing step 2.0
}
