% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adcpToolbox.R, R/readMeta.R
\name{read.meta}
\alias{read.meta}
\alias{read.meta}
\title{ADCP process step 2.1}
\usage{
read.meta(file, obj)

read.meta(file, obj)
}
\arguments{
\item{file}{csv file name}

\item{obj}{adp oce object to assign metadata to}

\item{file}{csv file name}

\item{obj}{adp oce object to assign metadata to}
}
\description{
Read in all metadata from csv template to adp object, sourced from log sheets

Read in all metadata from csv template to adp object, sourced from log sheets
}
